+++ changelog.md
@@ -1,4 +1,8 @@
 # ACEmulator Change Log
+### 2018-08-12
+[OptimShi]
+* Added handling of Bool.NpcLooksLikeObject to the AppraiseInfo event.
+
 ### 2018-08-11
 [mcreedjr]
 * Added missing comma in updated ShardBase.sql to allow creation of updated Shard DB after Mag-nus' previous PR